moved fetchUserTypes to redux
